@@ -9,6 +9,9 @@
     "ant design"
   ],
   "private": true,
+  "engines": {
+    "node": ">=16"
+  },
   "homepage": ".",
   "bugs": {
     "url": "https://github.com/krwu/ServerStatus-web/issues"
@@ -21,28 +24,23 @@
   },
   "dependencies": {
     "@ant-design/icons": "^4.6.4",
-    "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@types/jest": "^26.0.15",
-    "@types/node": "^12.0.0",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
+    "@babel/core": ">=7.0.0-0 <8.0.0",
+    "@babel/plugin-syntax-flow": "^7.14.5",
+    "@babel/plugin-transform-react-jsx": "^7.14.9",
+    "@types/node": "16.0.0",
+    "@types/react": "18",
+    "@types/react-dom": "18",
     "antd": "^4.0.0",
     "babel-plugin-import": "^1.13.1",
-    "customize-cra": "^1.0.0",
-    "react": "^17.0.2",
-    "react-app-rewired": "^2.1.6",
-    "react-dom": "^17.0.2",
-    "react-intl-universal": "^2.4.2",
-    "react-scripts": "4.0.3",
-    "typescript": "^4.1.2",
-    "web-vitals": "^1.0.1"
+    "react": "18",
+    "react-dom": "18",
+    "react-intl-universal": "^2.5.3",
+    "react-scripts": "^5.0.1",
+    "typescript": "^4.1.2"
   },
   "scripts": {
-    "start": "react-app-rewired start",
-    "build": "react-app-rewired build",
-    "test": "react-app-rewired test",
+    "start": "react-scripts start",
+    "build": "react-scripts build",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
@@ -53,7 +51,7 @@
   },
   "browserslist": {
     "production": [
-      ">0.2%",
+      ">5%",
       "not dead",
       "not op_mini all"
     ],
@@ -66,6 +64,6 @@
   "devDependencies": {
     "@semantic-release/changelog": "^6.0.0",
     "@semantic-release/git": "^10.0.0",
-    "semantic-release": "^18.0.0"
+    "semantic-release": "^19.0.3"
   }
 }
